// src/api/model/Task.ts
/**
 * Generated by orval v7.19.0 🍺
 * Do not edit manually.
 * Redfish
 * This contains the definition of a Redfish service.
 * OpenAPI spec version: 2025.4
 */
import type { OdataV4Context } from './OdataV4Context';
import type { OdataV4Etag } from './OdataV4Etag';
import type { OdataV4Id } from './OdataV4Id';
import type { OdataV4Type } from './OdataV4Type';
import type { TaskActions } from './TaskActions';
import type { ResourceDescription } from './ResourceDescription';
import type { ResourceId } from './ResourceId';
import type { TaskLinks } from './TaskLinks';
import type { Message } from './Message';
import type { ResourceName } from './ResourceName';
import type { ResourceOem } from './ResourceOem';
import type { TaskPayload } from './TaskPayload';
import type { OdataV4IdRef } from './OdataV4IdRef';
import type { TaskState } from './TaskState';
import type { ResourceHealth } from './ResourceHealth';

/**
 * The `Task` schema contains information about a task that the Redfish task service schedules or executes.  Tasks represent operations that take more time than a client typically wants to wait.
 * @version 1.7.4
 */
export interface Task {
    '@odata.context'?: OdataV4Context;
    '@odata.etag'?: OdataV4Etag;
    '@odata.id': OdataV4Id;
    '@odata.type': OdataV4Type;
    /** The available actions for this resource. */
    Actions?: TaskActions;
    /** @nullable */
    readonly Description?: ResourceDescription;
    /** The date and time when the task was completed.  This property will only appear when the task is complete. */
    readonly EndTime?: string;
    /**
     * The estimated total time required to complete the task.
     * @nullable
     * @pattern ^P(\d+D)?(T(\d+H)?(\d+M)?(\d+(.\d+)?S)?)?$
     */
    readonly EstimatedDuration?: string | null;
    /** An indication of whether the contents of the payload are hidden from view after the task has been created.  If `true`, responses do not return the payload.  If `false`, responses return the payload.  If this property is not present when the task is created, the default is `false`. */
    readonly HidePayload?: boolean;
    readonly Id: ResourceId;
    /** Contains references to other resources that are related to this resource. */
    Links?: TaskLinks;
    /** An array of messages associated with the task. */
    Messages?: Message[];
    readonly Name: ResourceName;
    /** The OEM extension property. */
    Oem?: ResourceOem;
    /** The HTTP and JSON request payload details for this task, unless they are hidden from view by the service. */
    Payload?: TaskPayload;
    /**
     * The completion percentage of this task.
     * @minimum 0
     * @maximum 100
     * @nullable
     */
    readonly PercentComplete?: number | null;
    /** The date and time when the task was started. */
    readonly StartTime?: string;
    /** The link to a collection of sub-tasks for this task. */
    readonly SubTasks?: OdataV4IdRef;
    /** The URI of the task monitor for this task. */
    readonly TaskMonitor?: string;
    /** The state of the task. */
    readonly TaskState?: TaskState;
    /** The completion status of the task. */
    readonly TaskStatus?: ResourceHealth;
}