// src/api/model/UpdateService.ts
export interface UpdateService {
    '@odata.context'?: OdataV4Context;
    '@odata.etag'?: OdataV4Etag;
    '@odata.id': OdataV4Id;
    '@odata.type': OdataV4Type;
    /** The available actions for this resource. */
    Actions?: UpdateServiceActions;
    /** The link to a collection of client identity certificates provided to the server referenced by the `ImageURI` parameter in `SimpleUpdate`. */
    readonly ClientCertificates?: OdataV4IdRef;
    /** @nullable */
    readonly Description?: ResourceDescription;
    /** An inventory of firmware. */
    readonly FirmwareInventory?: OdataV4IdRef;
    /**
     * The URI used to perform an HTTP or HTTPS push update to the update service.  The format of the message is vendor-specific.
     * @deprecated
     */
    readonly HttpPushUri?: string;
    /**
     * The options for `HttpPushUri`-provided software updates.
     * @deprecated
     */
    HttpPushUriOptions?: UpdateServiceHttpPushUriOptions;
    /**
     * An indication of whether a client has reserved the `HttpPushUriOptions` properties for software updates.
     * @deprecated
     * @nullable
     */
    HttpPushUriOptionsBusy?: boolean | null;
    /**
     * An array of URIs that indicate where to apply the update image.
     * @deprecated
     */
    HttpPushUriTargets?: (string | null)[];
    /**
     * An indication of whether any client has reserved the `HttpPushUriTargets` property.
     * @deprecated
     * @nullable
     */
    HttpPushUriTargetsBusy?: boolean | null;
    readonly Id: ResourceId;
    /** An inventory of locally stored images. */
    readonly LocalImageStore?: OdataV4IdRef;
    /**
     * The number of bytes currently available in the local image store.
     * @nullable
     */
    readonly LocalImageStoreAvailableCapacityBytes?: number | null;
    /**
     * The total number of bytes available to the local image store.
     * @nullable
     */
    readonly LocalImageStoreTotalCapacityBytes?: number | null;
    /**
     * The maximum size in bytes of the software update image that this service supports.
     * @nullable
     */
    readonly MaxImageSizeBytes?: number | null;
    /** The URI used to perform a Redfish Specification-defined multipart HTTP or HTTPS push update to the update service. */
    readonly MultipartHttpPushUri?: string;
    readonly Name: ResourceName;
    /** The OEM extension property. */
    Oem?: ResourceOem;
    /** A link to the public key that is used with the `SimpleUpdate` action for the key-based authentication.  The GenerateSSHIdentityKeyPair and RemoveSSHIdentityKeyPair are used to update the key for the `SimpleUpdate` action. */
    readonly PublicIdentitySSHKey?: OdataV4IdRef;
    /** The link to a collection of server certificates for the server referenced by the `ImageURI` parameter in SimpleUpdate. */
    readonly RemoteServerCertificates?: OdataV4IdRef;
    /** The link to a collection of keys that can be used to authenticate the server referenced by the `ImageURI` parameter in `SimpleUpdate`. */
    readonly RemoteServerSSHKeys?: OdataV4IdRef;
    /**
     * An indication of whether this service is enabled.
     * @nullable
     */
    ServiceEnabled?: boolean | null;
    /** An inventory of software. */
    readonly SoftwareInventory?: OdataV4IdRef;
    /** The status and health of the resource and its subordinate or dependent resources. */
    Status?: ResourceStatus;
    /** The image format types supported by the service. */
    readonly SupportedUpdateImageFormats?: readonly UpdateServiceSupportedUpdateImageFormatType[];
    /** A link to the capabilities of the update service. */
    readonly UpdateServiceCapabilities?: OdataV4IdRef;
    /**
     * An indication of whether the service will verify the certificate of the server referenced by the `ImageURI` parameter in `SimpleUpdate` prior to sending the transfer request.
     * @nullable
     */
    VerifyRemoteServerCertificate?: boolean | null;
    /**
     * An indication of whether the service will verify the SSH key of the server referenced by the `ImageURI` parameter in `SimpleUpdate` prior to sending the transfer request.
     * @nullable
     */
    VerifyRemoteServerSSHKey?: boolean | null;
}